+++ modules/hcsr04.h
@@ -7,11 +7,14 @@
 #define __HCSR04_H
 
 
+#include "common.h"
+
+
 #define HCSR04_TRIG     P2_1
 #define HCSR04_ECHO     P2_0
 
 
-unsigned int hcsr04_read(void);
+uint16_t hcsr04_read(void);
 
 
 #endif /* __HCSR04_H */
